requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
yt-dlp
ffmpeg-python
cloudinary